split out make_nested_schema and make_flat_schema text fixtures
