+++ crates/mongodb-connector/src/main.rs
@@ -7,6 +7,9 @@ mod query_context;
 mod query_response;
 mod schema;
 
+#[cfg(test)]
+mod test_helpers;
+
 use std::error::Error;
 
 use mongo_connector::MongoConnector;
